Добавил пакет `grpcapp` для работы с grpc-сервером

+++ internal/config/config.go
@@ -16,7 +16,7 @@ type Config struct {
 }
 
 type GRPCConfig struct {
-	Port    int           `yaml:"port"`
+	Port    int64         `yaml:"port"`
 	Timeout time.Duration `yaml:"timeout"`
 }
